Add files via upload
This problem is the 15th challenge out of a series of 5000 problems. It consists of two classes: the first is the Solution class, which contains the core implementation, and the second is for handling test cases. Additionally, this folder includes two images: one provides a description of the problem, and the other showcases the solution.

The problem involves calculating the power of a number raised to the power of its reversed value. It is categorized as a medium-difficulty problem and was sourced from the GeeksforGeeks platform. This problem helped me establish an efficient approach to solving exponentiation based on the problem's specific requirements.