Add files via upload
This problem is the 14th challenge out of a series of 5000 problems. It consists of two classes: the first is the Solution class, which contains the core implementation for reversing a string, and the second is for handling test cases to validate the functionality. Additionally, this folder includes two images: one provides a description of the problem, and the other showcases the solution.

The problem involves reversing the characters in a given string. It is categorized as an easy difficulty problem and was sourced from the GeekforGeeks platform. This problem helped me develop my understanding of string manipulation techniques and efficient usage of the StringBuilder class in Java